# perf: reduce memory usage by 20 percentage