Summary:

Change the TrieNode from hash-like table to array-like table to reduce memory usage.


Array-like table
```
RADIX_ROUTER_ROUTES=100000 RADIX_ROUTER_TIMES=10000000 luajit benchmark/simple-variable.lua
========== variable ==========
routes  :       100000
times   :       10000000
elapsed :       1.306459 s
QPS     :       7654277
ns/op   :       0.1306459 ns
path    :       /1/foo
handler :       1
Memory  :       110.10 MB
```

Hash-like table
```
RADIX_ROUTER_ROUTES=100000 RADIX_ROUTER_TIMES=10000000 luajit benchmark/simple-variable.lua
========== variable ==========
routes  :       100000
times   :       10000000
elapsed :       1.565496 s
QPS     :       6387751
ns/op   :       0.1565496 ns
path    :       /1/foo
handler :       1
Memory  :       141.62 MB

```